requirements.txt
scipy
matplotlib
opencv-python
moviepy